% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{clusterAssign}
\alias{clusterAssign}
\title{assign clusters to non-peak points}
\usage{
clusterAssign(peakID, higherID, rho)
}
\arguments{
\item{peakID}{A vector of the peak ID.}

\item{higherID}{A vector of IDs of the nearest neighbor with higer density.}

\item{rho}{A vector of the density values.}
}
\value{
the cluster ID
}
\description{
assign clusters to non-peak points
}

