% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{ClusterX}
\alias{ClusterX}
\title{Fast clustering by automaticly search and find of density peaks}
\usage{
ClusterX(data, dimReduction = NULL, outDim = 2, dc, gaussian = TRUE,
  alpha = 0.001, detectHalos = FALSE, SVMhalos = FALSE,
  parallel = FALSE, nCore = 4)
}
\arguments{
\item{data}{A data matrix for clustering.}

\item{dimReduction}{Dimenionality reduciton method.}

\item{outDim}{Number of dimensions will be used for clustering.}

\item{dc}{Distance cutoff value.}

\item{gaussian}{If apply gaussian to esitmate the density.}

\item{alpha}{Signance level for peak detection.}

\item{detectHalos}{If detect the halos.}

\item{SVMhalos}{If apply SVM model from cores to assign halos.}

\item{parallel}{If run the algorithm in parallel.}

\item{nCore}{Number of cores umployed for parallel compution.}
}
\value{
a object of \code{ClusterX} class
}
\description{
This package implement the clustering algorithm described by Alex Rodriguez
and Alessandro Laio (2014) with improvements of automatic peak detection and
parallel implementation
}
\details{
ClusterX works on low dimensional data analysis (Dimensionality less than 5).
If input data is of high diemnsional, t-SNE is conducted to reduce the dimensionality.
}
\examples{
dir <- system.file("extdata", package = "ClusterX")
r15 <- read.table(paste(dir, "R15.txt", sep = .Platform$file.sep), header = FALSE)
r15_c <- ClusterX(r15[,c(1,2)])
clusterPlot(r15_c)
densityPlot(r15_c)
peakPlot(r15_c)

d31 <- read.table(paste(dir, "D31.txt", sep = .Platform$file.sep), header = FALSE)
d31_c <- ClusterX(d31[,c(1,2)])
clusterPlot(d31_c)
densityPlot(d31_c)
peakPlot(d31_c)
}
\author{
Chen Hao
}

