% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{detect_anoms_sd}
\alias{detect_anoms_sd}
\title{Outlier detection}
\usage{
detect_anoms_sd(data, max_anoms = 0.1, alpha = 0.01, direction = "pos")
}
\arguments{
\item{data}{A vectors of boservations.}

\item{max_anoms}{Maximal percentile of anomalies.}

\item{alpha}{The level of statistical significance with which to accept or reject anomalies.}

\item{direction}{Directionality of the anomalies to be detected. Options are: 'pos' | 'neg' | 'both'.}
}
\value{
A vector containing indexes of the anomalies (outliers).
}
\description{
Using generialized ESD to detect outliers, iterate and remove point with ares higher than lamda
in a univariate data set assumed to come from a normally distributed population.
}

