% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{estimateDc}
\alias{estimateDc}
\title{Estimate the distance cutoff (density neighbourhood) from down-sampled data}
\usage{
estimateDc(data, sampleSize = 5000, neighborRateLow = 0.01,
  neighborRateHigh = 0.02)
}
\arguments{
\item{data}{Numeric matrix of data or data frame.}

\item{sampleSize}{The size of the down-sampled data.}

\item{neighborRateLow}{The lower bound of the neighbor rate (default 0.01).}

\item{neighborRateHigh}{The upper bound of the neighbor rate (default 0.15).}
}
\value{
A numeric value giving the estimated distance cutoff value.
}
\description{
This function estimate a distance cutoff value from the down-samples data,
wchich meet the criteria that the average neighbor rate (number of points
within the distance cutoff value) fall between the provided range.
}

