% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{haloDetect}
\alias{haloDetect}
\title{differentiate halo form cores}
\usage{
haloDetect(data, rho, cluster, peakID, dc)
}
\arguments{
\item{data}{Input data.}

\item{rho}{Density values.}

\item{cluster}{The assigned cluster labels.}

\item{peakID}{The peak IDs.}

\item{dc}{The distance cutoff value.}
}
\value{
the boolean value indicating if the point belongs to halo
}
\description{
differentiate halo form cores
}

