% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{peakDetect}
\alias{peakDetect}
\title{Automatic peak detection}
\usage{
peakDetect(rho, delta, alpha = 0.001)
}
\arguments{
\item{rho}{A vector of the local density, outout of \code{localDensity}}

\item{delta}{A vector of distance to closest observation of higher density}

\item{alpha}{The level of statistical significance for peak detection.}
}
\value{
a vector containing the indexes of peaks
}
\description{
Automatic detect peaks by searching high denisty point with anomalous large distance to
higher denisty peaks. rho and delta are transformed to one index, and the anomalous peaks
are detected using generalized ESD method.
}

